Changes that reaches certain accuracy

+++ SubjectPlanner.Core/Interfaces/ISubject.cs
@@ -3,7 +3,7 @@ namespace SubjectPlanner.Core;
 public interface ISubject
 {
     public IEnumerable<Schedule> Schedules { get; set; }
-    public float Hours { get; set; }
+    public double Hours { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
 }
